release 2.21.6 (#486)
* release 2.21.6 Hadar mainnet

+++ publish/src/commands/owner.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const fs = require('fs');
-const { black, gray, yellow, red, cyan, bgYellow } = require('chalk');
+const { gray, yellow, red, cyan, bgYellow, black } = require('chalk');
 const w3utils = require('web3-utils');
 const Web3 = require('web3');
 
@@ -15,7 +15,27 @@ const {
 	stringify,
 } = require('../util');
 
-const owner = async ({ network, newOwner, deploymentPath }) => {
+const {
+	getSafeInstance,
+	getSafeNonce,
+	getSafeTransactions,
+	checkExistingPendingTx,
+	createAndSaveApprovalTransaction,
+} = require('../safe-utils');
+
+const DEFAULTS = {
+	gasPrice: '15',
+	gasLimit: 2e5, // 200,000
+};
+
+const owner = async ({
+	network,
+	newOwner,
+	deploymentPath,
+	gasPrice = DEFAULTS.gasPrice,
+	gasLimit = DEFAULTS.gasLimit,
+	privateKey,
+}) => {
 	ensureNetwork(network);
 
 	if (!newOwner || !w3utils.isAddress(newOwner)) {
@@ -30,15 +50,35 @@ const owner = async ({ network, newOwner, deploymentPath }) => {
 		network,
 	});
 
-	const { providerUrl, etherscanLinkPrefix } = loadConnections({ network });
+	const { providerUrl, privateKey: envPrivateKey } = loadConnections({
+		network,
+	});
+
+	if (!privateKey) {
+		privateKey = envPrivateKey;
+	}
+
 	const web3 = new Web3(new Web3.providers.HttpProvider(providerUrl));
+	web3.eth.accounts.wallet.add(privateKey);
+	const account = web3.eth.accounts.wallet[0].address;
+	console.log(gray(`Using account with public key ${account}`));
+	console.log(gray(`Gas Price: ${gasPrice} gwei`));
+
+	let lastNonce;
+	// new owner should be gnosis safe proxy address
+	const protocolDaoContract = getSafeInstance(web3, newOwner);
+
+	// get protocolDAO nonce
+	const currentSafeNonce = await getSafeNonce(protocolDaoContract);
+
+	console.log(yellow(`Using Protocol DAO Safe contract at ${protocolDaoContract.options.address}`));
 
 	const confirmOrEnd = async message => {
 		try {
 			await confirmAction(
 				message +
 					cyan(
-						'\nPlease type "y" when transaction completed, or enter "n" to cancel and resume this later? (y/n) '
+						'\nPlease type "y" to stage transaction, or enter "n" to cancel and resume this later? (y/n) '
 					)
 			);
 		} catch (err) {
@@ -47,24 +87,57 @@ const owner = async ({ network, newOwner, deploymentPath }) => {
 		}
 	};
 
+	// Load staged transactions
+	const stagedTransactions = await getSafeTransactions({
+		network,
+		safeAddress: protocolDaoContract.options.address,
+	});
+
 	console.log(
 		gray('Running through operations during deployment that couldnt complete as not owner.')
 	);
-
+	// Read owner-actions.json + encoded data to stage tx's
 	for (const [key, entry] of Object.entries(ownerActions)) {
-		const { action, link, complete } = entry;
+		const { target, data, complete } = entry;
 		if (complete) continue;
 
-		await confirmOrEnd(
-			yellow('YOUR TASK: ') + `Invoke ${bgYellow(black(action))} (${key}) via ${cyan(link)}`
-		);
+		const existingTx = checkExistingPendingTx({
+			stagedTransactions,
+			target,
+			encodedData: data,
+			currentSafeNonce,
+		});
 
-		entry.complete = true;
-		fs.writeFileSync(ownerActionsFile, stringify(ownerActions));
+		if (existingTx) continue;
+
+		await confirmOrEnd(yellow('Confirm: ') + `Stage ${bgYellow(black(key))} to (${target})`);
+
+		try {
+			const newNonce = await createAndSaveApprovalTransaction({
+				safeContract: protocolDaoContract,
+				data,
+				to: target,
+				sender: account,
+				gasLimit,
+				gasPrice,
+				network,
+				lastNonce,
+			});
+
+			// track lastNonce submitted
+			lastNonce = newNonce;
+
+			entry.complete = true;
+			fs.writeFileSync(ownerActionsFile, stringify(ownerActions));
+		} catch (err) {
+			console.log(
+				gray(`Transaction failed, if sending txn to safe api failed retry manually - ${err}`)
+			);
+			return;
+		}
 	}
 
 	console.log(gray('Looking for contracts whose ownership we should accept'));
-
 	for (const contract of Object.keys(config)) {
 		const { address, source } = deployment.targets[contract];
 		const { abi } = deployment.sources[source];
@@ -80,11 +153,43 @@ const owner = async ({ network, newOwner, deploymentPath }) => {
 		if (currentOwner === newOwner) {
 			console.log(gray(`${newOwner} is already the owner of ${contract}`));
 		} else if (nominatedOwner === newOwner) {
-			await confirmOrEnd(
-				yellow(
-					`YOUR TASK: Invoke ${contract}.acceptOwnership() via ${etherscanLinkPrefix}/address/${address}#writeContract`
-				)
-			);
+			const encodedData = deployedContract.methods.acceptOwnership().encodeABI();
+
+			// Check if similar one already staged and pending
+			const existingTx = checkExistingPendingTx({
+				stagedTransactions,
+				target: deployedContract.options.address,
+				encodedData,
+				currentSafeNonce,
+			});
+
+			if (existingTx) continue;
+
+			// continue if no pending tx found
+			await confirmOrEnd(yellow(`Confirm: Stage ${contract}.acceptOwnership() via protocolDAO?`));
+
+			console.log(yellow(`Attempting action protocolDaoContract.approveHash()`));
+
+			try {
+				const newNonce = await createAndSaveApprovalTransaction({
+					safeContract: protocolDaoContract,
+					data: encodedData,
+					to: deployedContract.options.address,
+					sender: account,
+					gasLimit,
+					gasPrice,
+					network,
+					lastNonce,
+				});
+
+				// track lastNonce submitted
+				lastNonce = newNonce;
+			} catch (err) {
+				console.log(
+					gray(`Transaction failed, if sending txn to safe api failed retry manually - ${err}`)
+				);
+				return;
+			}
 		} else {
 			console.log(
 				cyan(
@@ -107,8 +212,11 @@ module.exports = {
 			)
 			.option(
 				'-o, --new-owner <value>',
-				'The address of you as owner (please include the 0x prefix)'
+				'The address of protocolDAO proxy contract as owner (please include the 0x prefix)'
 			)
+			.option('-v, --private-key [value]', 'The private key of wallet to stage with.')
+			.option('-g, --gas-price <value>', 'Gas price in GWEI', DEFAULTS.gasPrice)
+			.option('-l, --gas-limit <value>', 'Gas limit', parseInt, DEFAULTS.gasLimit)
 			.option('-n, --network <value>', 'The network to run off.', x => x.toLowerCase(), 'kovan')
 			.action(owner),
 };
